+++ src/store/setting/selectors/config.ts
@@ -1,3 +1,5 @@
+import { t } from 'i18next';
+
 import { SettingStore } from '@/store/setting';
 import { GenderEnum } from '@/types/agent';
 import { OpenAIConfig } from '@/types/config';
@@ -7,21 +9,16 @@ const currentOpenAIConfig = (s: SettingStore): OpenAIConfig => {
   return s.config.languageModel.openAI;
 };
 
-const currentFemaleTouchConfig = (s: SettingStore, touchArea: TouchAreaEnum): TouchAction[] => {
-  return s.config.touch[GenderEnum.FEMALE]?.[touchArea] || [];
-};
-
-const currentMaleTouchConfig = (s: SettingStore, touchArea: TouchAreaEnum): TouchAction[] => {
-  return s.config.touch[GenderEnum.MALE]?.[touchArea] || [];
-};
-
-const currentOtherTouchConfig = (s: SettingStore, touchArea: TouchAreaEnum): TouchAction[] => {
-  return s.config.touch[GenderEnum.OTHER]?.[touchArea] || [];
+const currentTouchConfig = (
+  s: SettingStore,
+  gender: GenderEnum,
+  touchArea: TouchAreaEnum,
+): TouchAction[] => {
+  const items = s.config.touch[gender]?.[touchArea] || [];
+  return items.map((item) => ({ ...item, text: t(item.text, { ns: 'constants' }) || item.text }));
 };
 
 export const configSelectors = {
   currentOpenAIConfig,
-  currentFemaleTouchConfig,
-  currentMaleTouchConfig,
-  currentOtherTouchConfig,
+  currentTouchConfig,
 };
